update variables (#42)
* delete kotlin

* define API-based methods for blockchains

* delete method

* put method

* APi-based methods of ledger account

* fixed API url

* remove explicitly importing

* change class to LedgerAccount

* 20 seconds timeout

* change ledger account name

* fix bcash

* ledger customer

* order book

* add Record

* ledger Subscription

* ledger Transaction

* lombok compile

* API-bases methods for KMS

* lombok compile

* applied lombok

* add details for blockchain API, test code

* update ApiKey and BaseUrl.java

* update API details and test code

* update code

* update details for API

* update details for API

* add details to APIs and test code

* remove unused import package

* Ledger account test

* test create account for the ledger

* test create account for the ledger

* unit test ledger account

* unit test customer ledger

* unit test customer ledger

* unit test order book

* unit test

* unit test

* clean code

* validate objects

* validate objects

* validate CreateAccount

* validate CreateAccount

* validate object and unit test

* clean code

* transaction for bitcoin and litecoin

* apply async

* apply async

* change names

* sign KMS transaction

* bcach transaction

* fix litecoin

* fix errors

* xrp transaction

* xrp with kms transaction

* send xrp transaction

* xrp wallet

* btc wallet

* change test package

* bip32

* derivation path

* refactor code

* refactor code

* fix input parameter

* don not use singleton objects to prevent from risk of race condition

* clean code

* generate private keys and addresses

* ethe wallet

* ethe wallet

* bch, vet wallets

* eth private key

* eth transaction

* create specific networks

* eth transaction

* eth transaction

* eth transaction

* eth transaction test

* Xlm transaction and eth transaction test

* offchain btc

* clean code

* eth offchain

* eth offchain

* ltc offchain

* xlm offchain

* update README.md

* update README.md

* generate Javadocs

* generate Javadocs

* xrp offchain

* fixed errors

* fixed errors

* apidocs

* kms offchain

* update Javadocs

* update Javadocs

* update Javadocs

* delete apidocs

* add java docs

* update encode and decode transaaction

* update with bitcoincashj

* bch offchain

* bcash address

* bcash offchain

* vet Transaction

* review code

* review code

* log message

* update logic as tatumjs

* bcash transaction

* update

* update variables

* testing

* update variables

* fixed errors

* fixed errors

* update document

* update variables

* refactor

* refactor

* refactor

* bech32 for change address

* refactor

* hide methods

* hide methods

* hide methods

* get API key and URL from environment

* offchain bitcoin test

* TATUM_API_KEY

* fixed an error

* fixed difficulty field

* fixed difficulty field

* fixed errors

* update test value

* fixed errors

* update doc

* update variables

* update variables

* update validator

+++ src/main/java/io/tatum/blockchain/XRP.java
@@ -99,7 +99,7 @@ public class XRP {
      * @throws ExecutionException   the execution exception
      * @throws InterruptedException the interrupted exception
      */
-    public String xrpGetLedger(BigDecimal ledgerIndex) throws ExecutionException, InterruptedException {
+    public String xrpGetLedger(BigInteger ledgerIndex) throws ExecutionException, InterruptedException {
         String uri = BaseUrl.getInstance().getUrl() + "/v3/xrp/ledger/" + ledgerIndex;
         return Async.get(uri);
     }
@@ -145,7 +145,7 @@ public class XRP {
      * @throws ExecutionException   the execution exception
      * @throws InterruptedException the interrupted exception
      */
-    public String xrpGetAccountTransactions(String address, BigDecimal min, String marker) throws ExecutionException, InterruptedException {
+    public String xrpGetAccountTransactions(String address, BigInteger min, String marker) throws ExecutionException, InterruptedException {
         String uri = BaseUrl.getInstance().getUrl() + "/v3/xrp/account/tx/" + address + "?min=" + min + "&marker=" + URLEncoder.encode(marker, StandardCharsets.UTF_8);
         System.out.println(uri);
         return Async.get(uri);
